Move enemy classes to enemy package
